schedule-builder/requirements.txt
flask>=3.0.0
google-generativeai>=0.8.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
